requirements.txt
fastapi
uvicorn
motor
pydantic
python-dotenv
langchain-community 
pymongo 
langchain-core 
langchain
openai
pypdf
pinecone
uuid
passlib
bcrypt 
python-jose[cryptography]
